session-3/Chinhook/Chinhook/requirements.txt
langchain
langchain_google_genai
python-dotenv
langchain-core
pandas
sqlalchemy
psycopg2-binary
openai
streamlit
google-generativeai


